profile id page ready + moving sign up component into page
