+++ our-people/src/.env
@@ -1,0 +1,1 @@
+VITE_API_URL=https://our-people-group-project.fly.dev/
