+++ our-people/src/pages/ProfileId.jsx
@@ -1,9 +1,55 @@
 import React, { useState, useEffect } from 'react'
 import ProjectCard from '../components/ProjectCard'
-import ProfileId from '../components/ProfileId'
 
 function ProfilePage() {
-    return ( <ProfileId />, <ProjectCard />
+    const [ profileData, setProfileData] = useState({profiles: []})
+    const { id } = useParams();
+
+    useEffect(() => {
+        fetch(`${import.meta.env.VITE_API_URL}profiles/${id}`)
+        .then((profileResults) => {
+            return profileResults.json();
+        })
+        .then((profileData) => {
+            setProfileData(profileData)
+        })
+    }, [])
+
+
+    const [projectList, setProjectList] = useState([]);
+
+    useEffect( () => {
+        fetch(`${import.meta.env.VITE_API_URL}projects`)
+        .then ((results) => {
+            return results.json();
+        })
+        .then ((data) => {
+            setProjectList(data)
+        });
+    }, []);
+
+    return ( 
+        <div>
+
+            <div>
+                <img src={profileData.profile_image}></img>
+                <h2>{profileData.display_name}</h2>
+                <h3>{profileData.pronouns}</h3>
+                <h3>{profileData.current_role}</h3>
+                <h3>{profileData.previous_role}</h3>
+                <p>{profileData.bio}</p>
+                <p>{profileData.challenges}</p>
+                <a>{profileData.github_url}</a>
+                <a>{profileData.linkedin_url}</a>
+            </div>
+
+            <div>
+            {projectList.map((projectData, key) => {
+                return <ProjectCard projectData={projectData}/>
+            })}
+            </div>
+
+        </div>
     )
 }
